Refactor OpenAI
