+++ council/llm/providers/__init__.py
@@ -2,3 +2,4 @@ from .anthropic import AnthropicLLM, AnthropicLLMConfiguration
 from .gemini import GeminiLLM, GeminiLLMConfiguration
 from .groq import GroqLLM, GroqLLMConfiguration
 from .ollama import OllamaLLM, OllamaLLMConfiguration
+from .openai import AzureLLM, AzureChatGPTConfiguration, OpenAILLM, OpenAIChatGPTConfiguration
