+++ council/llm/providers/openai/openai_chat_gpt_configuration.py
@@ -2,8 +2,7 @@ from __future__ import annotations
 
 from typing import Any, Final, Optional
 
-from council.llm import ChatGPTConfigurationBase
-from council.llm.llm_config_object import LLMConfigSpec
+from council.llm import LLMConfigSpec
 from council.utils import (
     Parameter,
     greater_than_validator,
@@ -12,6 +11,8 @@ from council.utils import (
     read_env_int,
     read_env_str,
 )
+
+from .chat_gpt_configuration import ChatGPTConfigurationBase
 
 _env_var_prefix: Final[str] = "OPENAI_"
