+++ council/llm/providers/openai/openai_llm.py
@@ -3,11 +3,12 @@ from __future__ import annotations
 from typing import Any, Optional
 
 import httpx
+from council.llm import LLMCallException, LLMCallTimeoutException, LLMConfigObject, LLMProviders
 from httpx import HTTPStatusError, TimeoutException
 
-from . import LLMCallException, LLMCallTimeoutException, OpenAIChatCompletionsModel, OpenAITokenCounter
-from .llm_config_object import LLMConfigObject, LLMProviders
+from .openai_chat_completions_llm import OpenAIChatCompletionsModel
 from .openai_chat_gpt_configuration import OpenAIChatGPTConfiguration
+from .openai_token_counter import OpenAITokenCounter
 
 
 class OpenAIChatCompletionsModelProvider:
