+++ tests/unit/llm/test_openai_token_counter.py
@@ -1,6 +1,7 @@
 import unittest
 
-from council.llm import LLMMessage, OpenAITokenCounter, LLMTokenLimitException
+from council.llm import LLMMessage, LLMTokenLimitException
+from council.llm.providers.openai.openai_token_counter import OpenAITokenCounter
 
 
 class TestLlmOpenAI(unittest.TestCase):
